Connected everything
Updated models to gui controllers and views
